二维码适配不同的服务器地址

+++ src/main/java/com/msgc/utils/qrCode/QrCodeUtilAdapter.java
@@ -12,9 +12,11 @@ import java.io.IOException;
 public class QrCodeUtilAdapter {
     //生成表分享的二维码，并缓存在本地
     public static void generate(Integer tableId) throws IOException, WriterException {
-        //生成路径
-        String outputPath = FilePath.getTableProcessedPath(tableId) + FilePath.FILE_NAME_TABLE_QrCODE_JPEG;
+        String outputPath = FilePath.getQrCodeFilePath(tableId);
         File jpegQrCode = new File(outputPath);
+        // 使用缓存
+        if(jpegQrCode.exists())
+            return;
         FileUtil.ensureExist(jpegQrCode);
         QrCodeUtil.createQrCode(new FileOutputStream(jpegQrCode), TableController.getShareCollectURL(tableId));
